Employee revision

+++ iH.Domain.Entities/Employee/Department.cs
@@ -1,8 +1,10 @@
 ﻿namespace iH.Domain.Employee.Entities
 {
+    using System;
+
     public class Department
     {
-        public int DepartmentId { get; set; }
+        public Int64 DepartmentId { get; set; }
         public string DepartmentName { get; set; }
     }
 }
