+++ iH.Domain.Entities/Employee/Designation.cs
@@ -1,8 +1,10 @@
 ﻿namespace iH.Domain.Employee.Entities
 {
+    using System;
+
     public class Designation
     {
-        public int DesignationId { get; set; }
+        public Int64 DesignationId { get; set; }
         public string DesignationName { get; set; }
     }
 }
