+++ iH.Domain.Entities/Employee/Location.cs
@@ -1,8 +1,10 @@
 ﻿namespace iH.Domain.Employee.Entities
 {
+    using System;
+
     public class Location
     {
-        public int LocationId { get; set; }
+        public Int64 LocationId { get; set; }
         public string LocationName { get; set; }
     }
 }
